Merge Feature/db into develop

+++ Api-restaurant/Dto/ArticleItemDTO.cs
@@ -6,7 +6,7 @@ namespace Api_restaurant.Dto
     {
         public int Id { get; set; }
         public string? Nom { get; set; }
-        public int Prix { get; set; }
+        public decimal Prix { get; set; }
         public string? Categorie { get; set; }
 
         public ArticleItemDTO() { }
